implementação de AlunoService, AlunoRepository, AlunoTest

+++ UniversidadeXYZ.Dominio/Entidades/Aluno.cs
@@ -7,7 +7,7 @@ namespace UniversidadeXYZ.Dominio.Entidades
     public class Aluno : BaseEntity
     {
         public string Nome { get; set; }
-        public int CPF { get; set; }
+        public long CPF { get; set; }
         public int Telefone { get; set; }
         public string Logradouro { get; set; }
